BUGFIX: Top navigation Week 1 link
The top navigation had a link for Week 1. When clicked it would download a file named "Unkown" onto the user machine. This was caused by a stray file in the /docs/ root. This stray file has a decent amount of transcribed work that I didn't want to loose. This has been moved to a new "_ignoreForNow" folder and documented in the readme.